[refactoring] Decompose backend
