--- price-map/apps/backend/src/app/app.service.spec.ts
@@ -1,21 +1,0 @@
-import { Test } from '@nestjs/testing';
-
-import { AppService } from './app.service';
-
-describe('AppService', () => {
-  let service: AppService;
-
-  beforeAll(async () => {
-    const app = await Test.createTestingModule({
-      providers: [AppService],
-    }).compile();
-
-    service = app.get<AppService>(AppService);
-  });
-
-  describe('getData', () => {
-    it('should return "Hello world на бэке вот так уот"', () => {
-      expect(service.getData()).toEqual({ message: 'Hello world на бэке вот так уот' });
-    });
-  });
-});
