--- price-map/apps/backend/src/app/app.service.ts
@@ -1,8 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-@Injectable()
-export class AppService {
-  getData(): { message: string } {
-    return { message: 'Hello world на бэке вот так уот' };
-  }
-}
