+++ price-map/apps/backend/src/app/constants/jwt.constant.ts
@@ -1,0 +1,3 @@
+export const jwtConstant = {
+  secret: 'secret-jwt-key',
+};
