+++ price-map/apps/backend/src/app/constants/role-key.constant.ts
@@ -1,0 +1,1 @@
+export const roleKey = 'roles';
