--- price-map/apps/backend/src/app/controllers/cats.service.ts
@@ -1,8 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-@Injectable()
-export class CatsService {
-  getData(): string {
-    return 'hook';
-  }
-}
