+++ price-map/apps/backend/src/app/decorators/role.decorator.ts
@@ -1,5 +1,5 @@
 import { SetMetadata } from '@nestjs/common';
-import { Role } from '../modules/auth/models/enums';
+import { Role } from '@price-map/core/enums';
+import { roleKey } from '../constants';
 
-export const ROLES_KEY = 'roles';
-export const Roles = (...roles: Role[]) => SetMetadata(ROLES_KEY, roles);
+export const Roles = (...roles: Role[]) => SetMetadata(roleKey, roles);
