+++ price-map/apps/backend/src/app/gateways/index.ts
@@ -1,0 +1,1 @@
+export { AppGateway } from './app.gateway';
