--- price-map/apps/backend/src/app/guards/auth.guard.ts
@@ -1,16 +1,0 @@
-import { Injectable, CanActivate, ExecutionContext } from '@nestjs/common';
-import { Observable } from 'rxjs';
-
-@Injectable()
-export class AuthGuard implements CanActivate {
-  canActivate(
-    context: ExecutionContext,
-  ): boolean | Promise<boolean> | Observable<boolean> {
-    const request = context.switchToHttp().getRequest();
-    return this.validateRequest(request);
-  }
-
-  public validateRequest(request): boolean {
-    return false;
-  }
-}
