+++ price-map/apps/backend/src/app/guards/index.ts
@@ -1,0 +1,2 @@
+export { JwtAuthGuard } from './jwt.guard';
+export { RolesAuthGuard } from './roles.guard';
