--- price-map/apps/backend/src/app/middlewares/logger.middleware.ts
@@ -1,10 +1,0 @@
-import { Injectable, NestMiddleware } from '@nestjs/common';
-import { Request, Response, NextFunction } from 'express';
-
-@Injectable()
-export class LoggerMiddleware implements NestMiddleware {
-  use(req: Request, res: Response, next: NextFunction) {
-    console.log('Request...' + req.method);
-    next();
-  }
-}
