+++ price-map/apps/backend/src/app/modules/auth/auth.module.ts
@@ -3,23 +3,20 @@ import { UsersModule } from '../users/users.module';
 import { PassportModule } from '@nestjs/passport';
 import { JwtModule } from '@nestjs/jwt';
 import { AuthService } from './services';
-import { jwtConstants } from './models/constants';
-import { JwtStrategy, LocalStrategy } from './strategies';
 import { AuthGateway } from './gateways';
+import { jwtConstant } from '../../constants';
 
 @Module({
   imports: [
     UsersModule,
     PassportModule,
     JwtModule.register({
-      secret: jwtConstants.secret,
+      secret: jwtConstant.secret,
       signOptions: { expiresIn: '10h' },
     }),
   ],
   providers: [
     AuthService,
-    LocalStrategy,
-    JwtStrategy,
     AuthGateway
   ],
   exports: [AuthService],
