--- price-map/apps/backend/src/app/modules/auth/guards/index.ts
@@ -1,3 +1,0 @@
-export { JwtAuthGuard } from './jwt.guard';
-export { LocalAuthGuard } from './local.guard';
-export { RolesAuthGuard } from './roles.guard';
