--- price-map/apps/backend/src/app/modules/auth/guards/local.guard.ts
@@ -1,10 +1,0 @@
-import { ExecutionContext, Injectable } from '@nestjs/common';
-import { AuthGuard } from '@nestjs/passport';
-
-@Injectable()
-export class LocalAuthGuard extends AuthGuard('local') {
-  handleRequest(err: any, user: any, info: any, context: ExecutionContext, status?: any): any {
-    console.log(err, user, info, context, status);
-    return {};
-  }
-}
