--- price-map/apps/backend/src/app/modules/auth/models/constants/index.ts
@@ -1,1 +1,0 @@
-export { jwtConstants } from './jwt.constant';
