--- price-map/apps/backend/src/app/modules/auth/models/constants/jwt.constant.ts
@@ -1,3 +1,0 @@
-export const jwtConstants = {
-  secret: 'secretKey',
-};
