+++ price-map/apps/backend/src/app/modules/auth/services/auth.service.ts
@@ -9,19 +9,4 @@ export class AuthService {
     private jwtService: JwtService
   ) {}
 
-  async validateUser(nickname: string, password: string): Promise<any> {
-    const user = await this.usersService.getByNickname(nickname);
-    if (user && user.password === password) {
-      const { password, ...result } = user;
-      return result;
-    }
-    return null;
-  }
-
-  async login(user: any) {
-    const payload = { nickname: user.nickname, userId: user.userId, role: user.role };
-    return {
-      token: this.jwtService.sign(payload),
-    };
-  }
 }
