--- price-map/apps/backend/src/app/modules/auth/strategies/index.ts
@@ -1,2 +1,0 @@
-export { JwtStrategy } from './jwt.strategy';
-export { LocalStrategy } from './local.strategy';
