--- price-map/apps/backend/src/app/modules/auth/strategies/jwt.strategy.ts
@@ -1,19 +1,0 @@
-import { ExtractJwt, Strategy } from 'passport-jwt';
-import { PassportStrategy } from '@nestjs/passport';
-import { Injectable } from '@nestjs/common';
-import { jwtConstants } from '../models/constants';
-
-@Injectable()
-export class JwtStrategy extends PassportStrategy(Strategy) {
-  constructor() {
-    super({
-      jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
-      ignoreExpiration: false,
-      secretOrKey: jwtConstants.secret
-    });
-  }
-
-  async validate(payload: any) {
-    return { userId: payload.userId, nickname: payload.nickname, role: payload.role };
-  }
-}
