--- price-map/apps/backend/src/app/modules/auth/strategies/local.strategy.ts
@@ -1,26 +1,0 @@
-import { Strategy } from 'passport-local';
-import { PassportStrategy } from '@nestjs/passport';
-import { Injectable, UnauthorizedException } from '@nestjs/common';
-import { AuthService } from '../services';
-
-@Injectable()
-export class LocalStrategy extends PassportStrategy(Strategy) {
-  constructor(private authService: AuthService) {
-    super({
-      usernameField: 'nickname'
-    });
-  }
-
-  async validate(nickname: string, password: string): Promise<any> {
-    console.log('validate', nickname, password);
-    const user = await this.authService.validateUser(nickname, password);
-    if (!user) {
-      throw new UnauthorizedException({
-        error: true,
-        message: 'Unauthorized',
-        statusCode: 401,
-      });
-    }
-    return user;
-  }
-}
