--- price-map/apps/backend/src/app/pipes/validation.pipe.ts
@@ -1,11 +1,0 @@
-import { PipeTransform, Injectable, ArgumentMetadata, BadRequestException } from '@nestjs/common';
-
-@Injectable()
-export class JustPipe implements PipeTransform {
-  transform(value: number, metadata: ArgumentMetadata) {
-    if (value > 100) {
-      throw new BadRequestException('Validation failed');
-    }
-    return value;
-  }
-}
